* ershov:
  Extract duplicate code in ParseManager to helper method
  Modify GameMode screen layout for iPad and iPhone, all orientations
  Fixed ParseManaged recreation
  Make Database singleton
  Add Pods to project, replaced FMDb with pod
  Adjust divider line size for different layouts
  Add font size adjust for iPad
  Improved main screen auto layout to adjust for orientation
  Fix compilation warnings.

Conflicts:
	src/Erundopel/Erundopel.xcodeproj/project.pbxproj
	src/Erundopel/Erundopel/GameVC.m
	src/Erundopel/Erundopel/MainMenuVC.m

+++ src/Erundopel/Erundopel/Card.m
@@ -20,7 +20,7 @@
 }
 
 - (void)shuffleMeanings {
-    int oldRightMeaningIndex = self.rightMeaningIndex;
+    unsigned int oldRightMeaningIndex = (unsigned int)self.rightMeaningIndex;
     
     self.rightMeaningIndex = arc4random() % 3;
     int falseMeaningIndexes[] = {
